Moved files: AllergensSection.js, CollapsibleSection.js, EditableRecipeDetails.js, and removed Header.js; updated imports in page.js
