+++ app/Recipe/[id]/page.js
@@ -3,11 +3,11 @@ import Image from 'next/image';
 import BackButton from "../../components/ui/BackButton";
 import { fetchRecipeById } from '../../api';
 import ImageGallery from '../../components/ImageGallery';
-import CollapsibleSection from '../../components/recipe-detail/CollapsibleSection';
+import CollapsibleSection from '../../components/CollapsibleSection';
 import Loading from './loading';
 import EditableRecipeDetails from '../../components/EditableRecipeDetails';
 import ReviewsSection from '../../components/ReviewsSection';
-import AllergensSection from '../../components/recipe-detail/AllergensSection';
+import AllergensSection from '../../components/AllergensSection';
 import RecipeIngredientsSelector from '../../components/RecipeIngredientsSelector';
 import { ShoppingListProvider } from '../../context/ShoppingListContext';
 import DownloadButton from '../../components/DownLoadButton'; // Import the new component
