+++ app/components/AllergensSection.js
